refactor: extract radio-button and radio-group logic to mixins
