+++ packages/radio-group/src/vaadin-radio-button-mixin.d.ts
@@ -1,0 +1,50 @@
+/**
+ * @license
+ * Copyright (c) 2017 - 2023 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import type { Constructor } from '@open-wc/dedupe-mixin';
+import type { ActiveMixinClass } from '@vaadin/a11y-base/src/active-mixin.js';
+import type { DelegateFocusMixinClass } from '@vaadin/a11y-base/src/delegate-focus-mixin.js';
+import type { DisabledMixinClass } from '@vaadin/a11y-base/src/disabled-mixin.js';
+import type { FocusMixinClass } from '@vaadin/a11y-base/src/focus-mixin.js';
+import type { KeyboardMixinClass } from '@vaadin/a11y-base/src/keyboard-mixin.js';
+import type { DelegateStateMixinClass } from '@vaadin/component-base/src/delegate-state-mixin.js';
+import type { CheckedMixinClass } from '@vaadin/field-base/src/checked-mixin.js';
+import type { InputMixinClass } from '@vaadin/field-base/src/input-mixin.js';
+import type { LabelMixinClass } from '@vaadin/field-base/src/label-mixin.js';
+
+/**
+ * Fired when the `checked` property changes.
+ */
+export type RadioButtonCheckedChangedEvent = CustomEvent<{ value: boolean }>;
+
+export interface RadioButtonCustomEventMap {
+  'checked-changed': RadioButtonCheckedChangedEvent;
+}
+
+export interface RadioButtonEventMap extends HTMLElementEventMap, RadioButtonCustomEventMap {}
+
+/**
+ * A mixin providing common radio-button functionality.
+ */
+export declare function RadioButtonMixin<T extends Constructor<HTMLElement>>(
+  base: T,
+): Constructor<ActiveMixinClass> &
+  Constructor<CheckedMixinClass> &
+  Constructor<DelegateFocusMixinClass> &
+  Constructor<DelegateStateMixinClass> &
+  Constructor<DisabledMixinClass> &
+  Constructor<FocusMixinClass> &
+  Constructor<InputMixinClass> &
+  Constructor<KeyboardMixinClass> &
+  Constructor<LabelMixinClass> &
+  Constructor<RadioButtonMixinClass> &
+  T;
+
+export declare class RadioButtonMixinClass {
+  /**
+   * The name of the radio button.
+   */
+  name: string;
+}
